+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 [1]: https://pypi.org/project/demisto-sdk/#history
 
 ### 0.4.3
+* Fixed an issue where the incident and indicator field BC check failed.
 
 ### 0.4.2
 * Fixed an issue where validate failed on Windows.